--- fullstack-junit-support/src/main/java/com/hedera/fullstack/junit/support/annotation/core/FullStackSuite.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2023 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hedera.fullstack.junit.support.annotation.core;
-
-import com.hedera.fullstack.junit.support.annotation.application.ApplicationNodes;
-import com.hedera.fullstack.junit.support.annotation.flow.MaxTestExecutionTime;
-import com.hedera.fullstack.junit.support.extensions.TestSuiteInitializer;
-import java.lang.annotation.*;
-import java.util.concurrent.TimeUnit;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-@Inherited
-@Documented
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@ExtendWith({TestSuiteInitializer.class})
-@ApplicationNodes(4)
-@MaxTestExecutionTime(value = 45, unit = TimeUnit.MINUTES)
-public @interface FullStackSuite {}
