--- fullstack-junit-support/src/main/java/com/hedera/fullstack/junit/support/annotation/core/TestExecutorType.java
@@ -1,13 +1,0 @@
-package com.hedera.fullstack.junit.support.annotation.core;
-
-public enum TestExecutorType {
-    /**
-     *
-     */
-    JAVA_DIRECT,
-
-    /**
-     *
-     */
-    NODE_MGMT_TOOLS
-}
