+++ fullstack-junit-support/src/main/java/com/hedera/fullstack/junit/support/annotations/application/ConfigurationValue.java
@@ -1,0 +1,14 @@
+package com.hedera.fullstack.junit.support.annotations.application;
+
+import java.lang.annotation.*;
+
+@Inherited
+@Documented
+@Repeatable(PlatformConfiguration.class)
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface ConfigurationValue {
+    String name();
+
+    String value();
+}
