+++ fullstack-junit-support/src/main/java/com/hedera/fullstack/junit/support/annotations/core/TestExecutorType.java
@@ -1,0 +1,13 @@
+package com.hedera.fullstack.junit.support.annotations.core;
+
+public enum TestExecutorType {
+    /**
+     *
+     */
+    JAVA_DIRECT,
+
+    /**
+     *
+     */
+    NODE_MGMT_TOOLS
+}
