+++ fullstack-junit-support/src/main/java/com/hedera/fullstack/junit/support/annotations/flow/WaitForDuration.java
@@ -1,0 +1,14 @@
+package com.hedera.fullstack.junit.support.annotations.flow;
+
+import java.lang.annotation.*;
+import java.util.concurrent.TimeUnit;
+
+@Inherited
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface WaitForDuration {
+    int value();
+
+    TimeUnit unit() default TimeUnit.SECONDS;
+}
