+++ fullstack-junit-support/src/main/java/module-info.java
@@ -1,11 +1,11 @@
 module com.hedera.fullstack.junit.support {
     // Global Exports
-    exports com.hedera.fullstack.junit.support.annotation.application;
-    exports com.hedera.fullstack.junit.support.annotation.core;
-    exports com.hedera.fullstack.junit.support.annotation.flow;
-    exports com.hedera.fullstack.junit.support.annotation.resource;
-    exports com.hedera.fullstack.junit.support.annotation.services;
-    exports com.hedera.fullstack.junit.support.annotation.validation;
+    exports com.hedera.fullstack.junit.support.annotations.application;
+    exports com.hedera.fullstack.junit.support.annotations.core;
+    exports com.hedera.fullstack.junit.support.annotations.flow;
+    exports com.hedera.fullstack.junit.support.annotations.resource;
+    exports com.hedera.fullstack.junit.support.annotations.services;
+    exports com.hedera.fullstack.junit.support.annotations.validation;
     exports com.hedera.fullstack.junit.support.events.application;
 
     // Targeted Exports
